fix(frontend/cs/r1cs): cast hint input variable to big.Int
The input is cast to *big.Int when the hint function is being computed.
Cast the inputs to the hint function to *big.Int already when calling
NewHint for more stable type. This allows later to restrict the number
of types a hint input can take (e.g. during serializing).