perf: allocate less in test engine
At every API call utils.FromInterface method was called. This method always
allocates new *big.Int, which is very stressful for the runtime. Instead, we
first check if the variable is already *big.Int and return it otherwise. Only
if type assertion fails do we allocate new *big.Int.

Now, as we do not allocate every time, we had to change the API calls to
allocate a new result, so that we wouldnt mutate the passed in values.