feat: add n to 1 MUX and MAP (#475)
* feat: update gnark version to v0.8.0

* Add an n to 1 multiplexer

This commit adds a simple n to 1 multiplexer to the std lib. This multiplexer is based on non-deterministic advice.

* Improve tests for `multiplexer.go`

Old tests were dependant to outputs and were not able to test invalid inputs correctly.

* Refactor multiplexer test

* Add a key value lookup table

* Change some names

A lookup table usually uses indices and not keys, so we renamed `Lookup` to `Map`. We also renamed the `gadgets` package to `selector`.

* docs: add examples into documentation

* docs: add package documentation

* docs: describe hint functions

* docs: remove profiling example

* refactor: rename 'sel' to 'queryKey'

* docs: correct typos

---------

Co-authored-by: Gautam Botrel <gautam.botrel@gmail.com>
Co-authored-by: Ivo Kubjas <ivo.kubjas@consensys.net>