fix: allocate new variable in engine.MulAcc (#482)
* fix: allocate new variable in engine.MulAcc

We previously modified the result inline, but the behaviour may be inconsistent
with actual frontend/compiler. When variable is a Term or LinearExpression,
then

   b := a

copies value of a into b and after that these variables are independent. For
example, with MulAcc(b, x, y) we won't modify the variable a. But, as test
engine modified variable inline, then doing MulAcc(b, x, y) also changed value
of a (as they are just *big.Int). This lead to inconsistent results.

* perf: use mulacc in emulated mul

* fix: return newly allocated value