fix: enforce emulated element width also in AssetIsEqual header
Previously we didn't conditionally enforce the width in the method as it was
done in the methods we called. However, there is a short path for constants and
in that case the check was omitted. Add width enforcement to the header of the
method to ensure it is also enforce for const inputs.