feat: range check gadget (#472)
* feat: add external range checker interface

* feat: add optional FrontendType method to the builders

In range checking gadget we try to estimate the number of constraints given
different parameters. But for estimating we need to know the costs of
operations. And the costs of the operations depend on the way we arithmetize
the circuit.

Added an internal interface which allows to query the arithmetization method
and implement this in existing builders.

* feat: implement range checking

* feat: use range checking in field emulation

* test: update circuit statistics

* test: update stats

* test: update stats