feat: add a partition selector (#486)
* feat: add a partition selector

* fix: handle pivot points at the end of input

We want to handle the case when `Partition` needs to select the whole input.

* docs: add missing docs

* test: add more tests

* fix: reduce mask size

It seems that handling pivot points at the end of the array was not a good idea. That will introduce the overhead of one more nonlinear constraint, while in many cased it is not needed.

* docs: fix some comments

* fix: use the new hint registration

Also removes a todo about a failing test, since that test is not failing anymore.

* fix: handle errors when calling hints

* refactor: make hints unexported

* docs: add a missing doc

Signed-off-by: aybehrouz <behrouz_ayati@yahoo.com>

* refactor: use per file hint registration

* refactor: use one `init` per package

Apparently, Go runs all the init functions of a package when the package is imported. So, there is no point in having multiple init functions in different source files.

* refactor: add GetHints for getting all hints

* chore: remove empty line in import

* feat: add `Slice` function

This commit adds a `Slice` function which is a wrapper for a cascade of two `Partition` calls. It also makes StepMask unexported.

* feat: handle full array selection

Now we handle the case when pivotPosition is at the start or end of the input array.

Signed-off-by: aybehrouz <behrouz_ayati@yahoo.com>

---------

Signed-off-by: aybehrouz <behrouz_ayati@yahoo.com>
Co-authored-by: Ivo Kubjas <ivo.kubjas@consensys.net>