fix: count possible overflow due to subtraction padding
Previously when estimating the maximal overflow for subtraction we only
considered the possible overflow caused by the subtraction padding and the
subtrahend. But actually as the padding may overflow the minuend we have to
also consider it.

Due to this, we also had to decrease the maximal possible overflow by one as
modular reduction uses subtraction as a subroutine.