feat: use two power for emulated BW6-761 limb size
Even though if we use a bit wider limb we get one less limb then there is a
mismatch between the machine and non-native representations. For now change it
to match the machine sizes.

We are likely to reconsider the sizes when have updated the range checker
gadget to have limb sizes multible of the table size.