fix: subtraction padding computation
Previously when computing the next overflow after subtraction we took it as
max(a.overflow+1, b.overflow+1), but it isn't correct due to wrong assumptions.

Actually, the padding is always one larger than b.overflow (we compute it as
such). And the next overflow is one more than max(padding, a.overflow) as when
we add a and padding we may get an overflow bit.