fix: use subtraction with reduce in AssertIsEqual
To show that two field elements are equal, we instead show that the difference
of the field elements is a multiple of emulated modulus. However, for computing
the difference we used non-reducing version of subtraction to avoid infinite
cycles. With the new mulmod implementation the reducing versions do not call
AssertIsEqual anymore so the infinite cycles are averted. For some edge cases
the difference may overflow scalar field and solving may fail.